[Test] test with kornia FL

+++ losses/focal_loss.py
@@ -47,9 +47,9 @@ class FocalLoss(nn.Module):
         C = inputs.size(1)
         P = F.softmax(inputs)
 
-        class_mask = inputs.data.new(N, C).fill_(0)
+        class_mask = inputs.long().data.new(N, C).fill_(0)
         class_mask = Variable(class_mask)
-        ids = targets.view(-1, 1)
+        ids = targets.long().view(-1, 1)
         class_mask.scatter_(1, ids.data, 1.)
         # print(class_mask)
